[pkg/winperfcounter] Remove redundant internal package
This is a follow up refactoring of pkg/winperfcounter. After latest changes the package became a thin wrapper over "pkg/winperfcounters/internal/pdh". This change moves code from pkg/winperfcounters/internal/pdh to pkg/winperfcounters without any additional changes.
