+++ pkg/winperfcounters/win_perf_counters_notwindows.go
@@ -15,4 +15,4 @@
 //go:build !windows
 // +build !windows
 
-package pdh // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/winperfcounters/internal/pdh"
+package winperfcounters // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/winperfcounters"
